refactor(tests): move iceberg rest credential expansion tests to unit test module
Tests for expand_env_refs (rest.rs) were inline in src/. Relocate to
tests/unit/io/write/iceberg_rest.rs, matching the project's test layout.

Required making expand_env_refs and the rest module pub (#[doc(hidden)])
so the external test crate can import them — inline was the only option
for a fully-private fn.
